move inference inside package
